Reorganized game_server package structure
