+++ P5/src/edu/colostate/cs/cs414/p5/client_server/server/game_server/GameServer.java
@@ -3,6 +3,8 @@ package edu.colostate.cs.cs414.p5.client_server.server.game_server;
 import java.io.IOException;
 import java.net.InetSocketAddress;
 
+import edu.colostate.cs.cs414.p5.client_server.server.game_server.game_manager.GameManager;
+import edu.colostate.cs.cs414.p5.client_server.server.game_server.invite_manager.GameInviteManager;
 import edu.colostate.cs.cs414.p5.client_server.server.registry.AbstractRegistry;
 import edu.colostate.cs.cs414.p5.client_server.server.registry.ActiveRegistry;
 import edu.colostate.cs.cs414.p5.client_server.server.session.ClientSession;
