+++ P5/src/edu/colostate/cs/cs414/p5/client_server/server/session/SessionManager.java
@@ -12,8 +12,8 @@ import java.util.TimerTask;
 import edu.colostate.cs.cs414.p5.client_server.logger.Logger;
 import edu.colostate.cs.cs414.p5.client_server.server.AbstractServer;
 import edu.colostate.cs.cs414.p5.client_server.server.ActiveServer;
-import edu.colostate.cs.cs414.p5.client_server.server.game_server.GameInviteManager;
-import edu.colostate.cs.cs414.p5.client_server.server.game_server.GameManager;
+import edu.colostate.cs.cs414.p5.client_server.server.game_server.game_manager.GameManager;
+import edu.colostate.cs.cs414.p5.client_server.server.game_server.invite_manager.GameInviteManager;
 import edu.colostate.cs.cs414.p5.client_server.server.registry.AbstractRegistry;
 import edu.colostate.cs.cs414.p5.client_server.server.registry.ActiveRegistry;
 import edu.colostate.cs.cs414.p5.client_server.transmission.Task;
